+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 Added
 ******
 * bowtie2 and kallisto paired-end modes now support the new file naming method 'smart', that attempts to automatically determine the common name of each pair of paired-end fastq files.
+* Added a p-value histogram plot (DESeqFilter.pval_histogram) that displays the distribution of p-values in a differential expression table.
 
 Changed
 ********